+++ relay-server/src/actors/events.rs
@@ -682,6 +682,7 @@ impl EventProcessor {
 
             // session data is never considered as part of deduplication
             ItemType::Session => false,
+            ItemType::Sessions => false,
         }
     }