+++ CHANGELOG.md
@@ -1,12 +1,23 @@
 # Changelog
 
+## Unreleased
+
+**Features**:
+
+- Add dynamic sampling for transactions. ([#835](https://github.com/getsentry/relay/pull/835))
+
+**Bug Fixes**:
+
+
+**Internal**:
+
+
 ## 20.11.0
 
 **Features**:
 
 - Rename upstream retries histogram metric and add upstream requests duration metric. ([#816](https://github.com/getsentry/relay/pull/816))
 - Add options for metrics buffering (`metrics.buffering`) and sampling (`metrics.sample_rate`). ([#821](https://github.com/getsentry/relay/pull/821))
-- Add dynamic sampling for transactions. ([#835](https://github.com/getsentry/relay/pull/835))
 
 **Bug Fixes**: